Split AzureStackHciLoadBalancer code file
